INTERMEDIATE CHANGES
* separating unit tests and integration tests for gcs_utils, combine_ehr_rdr, and resources
* converting validation_test to it's unit and integration tests.
* Removing files that were moved to other locations:  combine_ehr_rdr_test, resources_test, validation_test
* splits unit and integration tests for bq_utils_test
* splits unit and integration tests for generate_ext_tables_tests
* separate unit and integration tests for retract_data_bq_test.py
* separating retract_data_gcs_test unit and integration tests
* change class name to SnapshotByQueryTest in snapshot_by_query_test
* pylint changes to completeness_test
