+++ data_steward/tests/integration_tests/data_steward/tools/combine_ehr_rdr_test.py
@@ -3,50 +3,27 @@ import logging
 import os
 import sys
 import unittest
+from io import open
 
 # Project imports
 import bq_utils
-import common
 import gcs_utils
 import resources
 from tests import test_util
 from constants.tools.combine_ehr_rdr import EHR_CONSENT_TABLE_ID, RDR_TABLES_TO_COPY, DOMAIN_TABLES
-from tools.combine_ehr_rdr import copy_rdr_table, ehr_consent, main, mapping_table_for, create_cdm_tables
-from tools.combine_ehr_rdr import logger
-from tools.combine_ehr_rdr import mapping_query
-
-EXPECTED_MAPPING_QUERY = ('SELECT DISTINCT'
-                          ' \'{rdr_dataset_id}\'  AS src_dataset_id,'
-                          '  {domain_table}_id  AS src_{domain_table}_id,'
-                          '  \'rdr\' as src_hpo_id,'
-                          '  {domain_table}_id + {mapping_constant}  AS {domain_table}_id'
-                          '  FROM {rdr_dataset_id}.{domain_table}'
-                          ''
-                          '  UNION ALL'
-                          ''
-                          '  SELECT DISTINCT'
-                          '  \'{ehr_dataset_id}\'  AS src_dataset_id,'
-                          '  t.{domain_table}_id AS src_{domain_table}_id,'
-                          '  v.src_hpo_id AS src_hpo_id,'
-                          '  t.{domain_table}_id  AS {domain_table}_id'
-                          '  FROM {ehr_dataset_id}.{domain_table} t'
-                          '  JOIN {ehr_dataset_id}._mapping_{domain_table}  v on t.{domain_table}_id = v.{'
-                          'domain_table}_id'
-                          '  WHERE EXISTS'
-                          '  (SELECT 1 FROM {combined_dataset_id}.{ehr_consent_table_id} c'
-                          '  WHERE t.person_id = c.person_id)')
+from tools.combine_ehr_rdr import (copy_rdr_table, ehr_consent, main, mapping_table_for,
+                                   create_cdm_tables, logger)
 
 UNCONSENTED_EHR_COUNTS_QUERY = ('  select \'{domain_table}\' as table_id, count(1) as n from (SELECT DISTINCT'
                                 '  v.src_hpo_id AS src_hpo_id,'
                                 '  t.{domain_table}_id  AS {domain_table}_id'
                                 '  FROM {ehr_dataset_id}.{domain_table} t'
-                                '  JOIN {ehr_dataset_id}._mapping_{domain_table}  v on t.{domain_table}_id = v.{'
-                                'domain_table}_id'
+                                '  JOIN {ehr_dataset_id}._mapping_{domain_table}  v'
+                                '  on t.{domain_table}_id = v.{domain_table}_id'
                                 '  WHERE NOT EXISTS'
                                 '  (SELECT 1 FROM {combined_dataset_id}.{ehr_consent_table_id} c'
                                 '  WHERE t.person_id = c.person_id))')
 
-@unittest.skipIf(os.getenv('ALL_TESTS') == 'False', 'Skip CombineEhrRdrTest cases')
 class CombineEhrRdrTest(unittest.TestCase):
 
     @classmethod
@@ -86,20 +63,20 @@ class CombineEhrRdrTest(unittest.TestCase):
         app_id = bq_utils.app_identity.get_application_id()
         filename = table + '.csv'
         schema = os.path.join(resources.fields_path, table + '.json')
-        f = os.path.join(path, filename)
-        if os.path.exists(os.path.join(path, filename)):
-            with open(f, 'rb') as fp:
-                gcs_utils.upload_object(bucket, filename, fp)
-        else:
+
+        file_path = os.path.join(path, filename)
+        try:
+            with open(file_path, 'rb') as filepath:
+                gcs_utils.upload_object(bucket, filename, filepath)
+        except OSError:
             test_util.write_cloud_str(bucket, filename, '\n')
+
         gcs_path = 'gs://{bucket}/{filename}'.format(bucket=bucket, filename=filename)
         load_results = bq_utils.load_csv(schema, gcs_path, app_id, dataset_id, table, allow_jagged_rows=True)
         load_job_id = load_results['jobReference']['jobId']
         return load_job_id
 
     def setUp(self):
-        super(CombineEhrRdrTest, self).setUp()
-        self.APP_ID = bq_utils.app_identity.get_application_id()
         self.ehr_dataset_id = bq_utils.get_dataset_id()
         self.rdr_dataset_id = bq_utils.get_rdr_dataset_id()
         self.combined_dataset_id = bq_utils.get_combined_dataset_id()
@@ -118,8 +95,13 @@ class CombineEhrRdrTest(unittest.TestCase):
          7: NULL and Yes with same date/time
         """
         # sanity check
+        # pre-conditions
         self.assertFalse(bq_utils.table_exists(EHR_CONSENT_TABLE_ID, self.combined_dataset_id))
+
+        # test
         ehr_consent()
+
+        # post conditions
         self.assertTrue(bq_utils.table_exists(EHR_CONSENT_TABLE_ID, self.combined_dataset_id),
                         'Table {dataset}.{table} created by consented_person'.format(dataset=self.combined_dataset_id,
                                                                                      table=EHR_CONSENT_TABLE_ID))
@@ -141,17 +123,21 @@ class CombineEhrRdrTest(unittest.TestCase):
             self.assertTrue(actual, msg='RDR table {table} should be copied'.format(table=table))
 
             # Check that row count in combined is same as rdr
-            q = '''
-              WITH rdr AS
-               (SELECT COUNT(1) n FROM {rdr_dataset_id}.{table}),
-              combined AS
-               (SELECT COUNT(1) n FROM {combined_dataset_id}.{table})
-              SELECT
-                rdr.n      AS rdr_count,
-                combined.n AS combined_count
-              FROM rdr, combined
-            '''.format(rdr_dataset_id=self.rdr_dataset_id, combined_dataset_id=self.combined_dataset_id, table=table)
-            response = bq_utils.query(q)
+            query = (
+                'WITH rdr AS '
+                ' (SELECT COUNT(1) n FROM {rdr_dataset_id}.{table}), '
+                'combined AS '
+                ' (SELECT COUNT(1) n FROM {combined_dataset_id}.{table}) '
+                'SELECT '
+                'rdr.n AS rdr_count, '
+                'combined.n AS combined_count '
+                'FROM rdr, combined '
+            ).format(
+                rdr_dataset_id=self.rdr_dataset_id,
+                combined_dataset_id=self.combined_dataset_id,
+                table=table
+            )
+            response = bq_utils.query(query)
             rows = bq_utils.response2rows(response)
             self.assertTrue(len(rows) == 1)  # sanity check
             row = rows[0]
@@ -160,41 +146,29 @@ class CombineEhrRdrTest(unittest.TestCase):
             self.assertEqual(rdr_count, combined_count,
                              msg_fmt.format(table=table, rdr_count=rdr_count, combined_count=combined_count))
 
-    def test_mapping_query(self):
-        table_name = 'visit_occurrence'
-        q = mapping_query(table_name)
-        expected_query = EXPECTED_MAPPING_QUERY.format(rdr_dataset_id=bq_utils.get_rdr_dataset_id(),
-                                                       ehr_dataset_id=bq_utils.get_dataset_id(),
-                                                       combined_dataset_id=self.combined_dataset_id,
-                                                       domain_table=table_name,
-                                                       mapping_constant=common.RDR_ID_CONSTANT,
-                                                       ehr_consent_table_id=EHR_CONSENT_TABLE_ID)
-
-        self.assertEqual(expected_query, q, "Mapping query for \n {q} \n to is not as expected".format(q=q))
-
     def _ehr_only_records_excluded(self):
         """
         EHR person records which are missing from RDR are excluded from combined
         """
-        q = '''
-        WITH ehr_only AS
-        (SELECT person_id
-         FROM {ehr_dataset_id}.person ep
-         WHERE NOT EXISTS
-           (SELECT 1
-            FROM {rdr_dataset_id}.person rp
-            WHERE rp.person_id = ep.person_id)
-        )
-        SELECT
-          ehr_only.person_id AS ehr_person_id,
-          p.person_id        AS combined_person_id
-        FROM ehr_only
-          LEFT JOIN {combined_dataset_id}.person p
-            ON ehr_only.person_id = p.person_id
-        '''.format(ehr_dataset_id=self.ehr_dataset_id,
-                   rdr_dataset_id=self.rdr_dataset_id,
-                   combined_dataset_id=self.combined_dataset_id)
-        response = bq_utils.query(q)
+        query = (
+            'WITH ehr_only AS '
+            ' (SELECT person_id '
+            '  FROM {ehr_dataset_id}.person ep '
+            '  WHERE NOT EXISTS '
+            '    (SELECT 1 '
+            '     FROM {rdr_dataset_id}.person rp '
+            '     WHERE rp.person_id = ep.person_id) '
+            ' ) '
+            'SELECT '
+            'ehr_only.person_id AS ehr_person_id, '
+            'p.person_id AS combined_person_id '
+            'FROM ehr_only '
+            'LEFT JOIN {ehr_rdr_dataset_id}.person p '
+            'ON ehr_only.person_id = p.person_id'
+        ).format(ehr_dataset_id=self.ehr_dataset_id,
+                 rdr_dataset_id=self.rdr_dataset_id,
+                 ehr_rdr_dataset_id=self.combined_dataset_id)
+        response = bq_utils.query(query)
         rows = bq_utils.response2rows(response)
         self.assertGreater(len(rows), 0, 'Test data is missing EHR-only records')
         for row in rows:
@@ -203,12 +177,14 @@ class CombineEhrRdrTest(unittest.TestCase):
                               'EHR-only person_id `{ehr_person_id}` found in combined when it should be excluded')
 
     def get_unconsented_ehr_records_count(self, table_name):
-        q = UNCONSENTED_EHR_COUNTS_QUERY.format(rdr_dataset_id=bq_utils.get_rdr_dataset_id(),
-                                                ehr_dataset_id=bq_utils.get_dataset_id(),
-                                                combined_dataset_id=self.combined_dataset_id,
-                                                domain_table=table_name,
-                                                ehr_consent_table_id='_ehr_consent')
-        response = bq_utils.query(q)
+        query = UNCONSENTED_EHR_COUNTS_QUERY.format(
+            rdr_dataset_id=bq_utils.get_rdr_dataset_id(),
+            ehr_dataset_id=bq_utils.get_dataset_id(),
+            ehr_rdr_dataset_id=self.combined_dataset_id,
+            domain_table=table_name,
+            ehr_consent_table_id='_ehr_consent'
+        )
+        response = bq_utils.query(query)
         rows = bq_utils.response2rows(response)
         return rows[0]['n']
 
@@ -216,21 +192,21 @@ class CombineEhrRdrTest(unittest.TestCase):
         """
         Check mapping tables exist, have correct schema, have expected number of records
         """
-        where = '''
-                WHERE EXISTS
-                   (SELECT 1 FROM {combined_dataset_id}.{ehr_consent_table_id} c
-                    WHERE t.person_id = c.person_id)
-                '''.format(combined_dataset_id=self.combined_dataset_id, ehr_consent_table_id=EHR_CONSENT_TABLE_ID)
+        where = (
+            'WHERE EXISTS '
+            '  (SELECT 1 FROM {ehr_rdr_dataset_id}.{ehr_consent_table_id} c '
+            '   WHERE t.person_id = c.person_id)'
+        ).format(ehr_rdr_dataset_id=self.combined_dataset_id,
+                 ehr_consent_table_id=EHR_CONSENT_TABLE_ID)
         ehr_counts = test_util.get_table_counts(self.ehr_dataset_id, DOMAIN_TABLES, where)
         rdr_counts = test_util.get_table_counts(self.rdr_dataset_id)
         combined_counts = test_util.get_table_counts(self.combined_dataset_id)
         output_tables = combined_counts.keys()
         expected_counts = dict()
         expected_diffs = ['observation']
-        self.maxDiff = None
 
-        for t in DOMAIN_TABLES:
-            expected_mapping_table = mapping_table_for(t)
+        for table in DOMAIN_TABLES:
+            expected_mapping_table = mapping_table_for(table)
             self.assertIn(expected_mapping_table, output_tables)
             expected_fields = resources.fields_for(expected_mapping_table)
             actual_table_info = bq_utils.get_table_info(expected_mapping_table, self.combined_dataset_id)
@@ -240,14 +216,19 @@ class CombineEhrRdrTest(unittest.TestCase):
 
             # Count should be sum of EHR and RDR
             # (except for tables like observation where extra records are created for demographics)
-            if 'person_id' in [field['name'] for field in resources.fields_for(t)]:
-                unconsented_ehr_records = self.get_unconsented_ehr_records_count(t)
+            if 'person_id' in [field.get('name', '') for field in resources.fields_for(table)]:
+                unconsented_ehr_records = self.get_unconsented_ehr_records_count(table)
             else:
                 unconsented_ehr_records = 0
+
             actual_count = combined_counts[expected_mapping_table]
-            expected_count = actual_count if t in expected_diffs \
-                else (ehr_counts[t] - unconsented_ehr_records) + rdr_counts[t]
+
+            if table in expected_diffs:
+                expected_count = actual_count
+            else:
+                expected_count = (ehr_counts[table] - unconsented_ehr_records) + rdr_counts[table]
             expected_counts[expected_mapping_table] = expected_count
+
         self.assertDictContainsSubset(expected_counts, combined_counts)
 
     def _all_rdr_records_included(self):
@@ -256,30 +237,36 @@ class CombineEhrRdrTest(unittest.TestCase):
         """
         for domain_table in DOMAIN_TABLES:
             mapping_table = mapping_table_for(domain_table)
-            q = '''SELECT rt.{domain_table}_id as id
-               FROM {rdr_dataset_id}.{domain_table} rt
-               LEFT JOIN {combined_dataset_id}.{mapping_table} m
-               ON rt.{domain_table}_id = m.src_{domain_table}_id
-               WHERE
-                 m.{domain_table}_id IS NULL
-               OR NOT EXISTS
-                 (SELECT 1 FROM {combined_dataset_id}.{domain_table} t
-                  WHERE t.{domain_table}_id = m.{domain_table}_id)'''.format(
-                domain_table=domain_table,
-                rdr_dataset_id=bq_utils.get_rdr_dataset_id(),
-                combined_dataset_id=bq_utils.get_combined_dataset_id(),
-                mapping_table=mapping_table)
-            response = bq_utils.query(q)
+            query = (
+                'SELECT rt.{domain_table}_id as id '
+                'FROM {rdr_dataset_id}.{domain_table} rt '
+                'LEFT JOIN {ehr_rdr_dataset_id}.{mapping_table} m '
+                'ON rt.{domain_table}_id = m.src_{domain_table}_id '
+                'WHERE '
+                '  m.{domain_table}_id IS NULL '
+                'OR NOT EXISTS '
+                ' (SELECT 1 FROM {ehr_rdr_dataset_id}.{domain_table} t '
+                '  WHERE t.{domain_table}_id = m.{domain_table}_id)').format(
+                    domain_table=domain_table,
+                    rdr_dataset_id=bq_utils.get_rdr_dataset_id(),
+                    ehr_rdr_dataset_id=bq_utils.get_ehr_rdr_dataset_id(),
+                    mapping_table=mapping_table)
+            response = bq_utils.query(query)
             rows = bq_utils.response2rows(response)
             self.assertEqual(0, len(rows), "RDR records should map to records in mapping and combined tables")
 
     def test_create_cdm_tables(self):
+        # pre-conditions
         # Sanity check
         tables_before = bq_utils.list_tables(self.combined_dataset_id)
         table_names_before = [t['tableReference']['tableId'] for t in tables_before]
         for table in resources.CDM_TABLES:
             self.assertNotIn(table, table_names_before)
+
+        # test
         create_cdm_tables()
+
+        # post conditions
         tables_after = bq_utils.list_tables(self.combined_dataset_id)
         table_names_after = [t['tableReference']['tableId'] for t in tables_after]
         for table in resources.CDM_TABLES:
@@ -292,7 +279,10 @@ class CombineEhrRdrTest(unittest.TestCase):
         pass
 
     def test_main(self):
+        # test
         main()
+
+        # post conditions
         self._mapping_table_checks()
         self._ehr_only_records_excluded()
         self._all_rdr_records_included()
