+++ data_steward/tests/unit_test/data_steward/tools/generate_ext_tables_test.py
@@ -1,13 +1,12 @@
 import unittest
+
 import mock
 
-import app_identity
-import bq_utils
-import tools.generate_ext_tables as gen_ext
 import common
-import resources
 import constants.bq_utils as bq_consts
 import constants.cdr_cleaner.clean_cdr as cdr_consts
+import resources
+import tools.generate_ext_tables as gen_ext
 
 
 class GenerateExtTablesTest(unittest.TestCase):
@@ -21,19 +20,20 @@ class GenerateExtTablesTest(unittest.TestCase):
     def setUp(self):
         self.project_id = 'project_id'
         self.dataset_id = 'dataset_id'
-        self.bq_project_id = app_identity.get_application_id()
-        self.bq_dataset_id = bq_utils.get_unioned_dataset_id()
-        self.obs_fields = [{
+        self.obs_fields = [
+            {
                 "type": "integer",
                 "name": "observation_id",
                 "mode": "nullable",
                 "description": "The observation_id used in the observation table."
-            }, {
+            },
+            {
                 "type": "string",
                 "name": "src_id",
                 "mode": "nullable",
                 "description": "The provenance of the data associated with the observation_id."
-            }]
+            }
+        ]
         self.hpo_list = resources.hpo_csv()
         self.mapping_tables = [gen_ext.MAPPING_PREFIX + cdm_table
                                for cdm_table in common.AOU_REQUIRED
@@ -41,16 +41,26 @@ class GenerateExtTablesTest(unittest.TestCase):
         self.bq_string = '("{hpo_name}", "EHR site nnn")'
 
     def test_get_obs_fields(self):
+        # pre-conditions
         table = common.OBSERVATION
         expected = self.obs_fields
+
+        # test
         actual = gen_ext.get_table_fields(table)
+
+        # post conditions
         self.assertCountEqual(expected, actual)
 
     def test_get_cdm_table_id(self):
+        # pre-conditions
         observation_table_id = common.OBSERVATION
         expected = observation_table_id
         mapping_observation = gen_ext.MAPPING_PREFIX + observation_table_id
+
+        # test
         actual = gen_ext.get_cdm_table_from_mapping(mapping_observation)
+
+        # post conditions
         self.assertCountEqual(expected, actual)
 
     def test_site_mapping_dict(self):
@@ -69,7 +79,7 @@ class GenerateExtTablesTest(unittest.TestCase):
         for hpo_item in hpo_rdr_mapping_list:
             self.assertIn(hpo_item[0], expected)
             if hpo_item[0] == gen_ext.RDR:
-                self.assertEquals(hpo_item[1], gen_ext.PPI_PM)
+                self.assertEqual(hpo_item[1], gen_ext.PPI_PM)
 
     def test_convert_to_bq_string(self):
         hpo_rdr_mapping_list = gen_ext.get_hpo_and_rdr_mappings()
@@ -79,13 +89,7 @@ class GenerateExtTablesTest(unittest.TestCase):
         hpo_bq_list.append('("{rdr}", "{ppi_pm}")'.format(rdr=gen_ext.RDR, ppi_pm=gen_ext.PPI_PM))
         expected = ', '.join(hpo_bq_list)
         actual = gen_ext.convert_to_bq_string(hpo_rdr_mapping_list)
-        self.assertEquals(len(actual), len(expected))
-
-    def test_create_populate_source_mapping_table(self):
-        mapping_list = gen_ext.get_hpo_and_rdr_mappings()
-        expected = str(len(mapping_list))
-        num_rows_affected = gen_ext.create_and_populate_source_mapping_table(self.bq_project_id, self.bq_dataset_id)
-        self.assertEquals(expected, num_rows_affected)
+        self.assertEqual(len(actual), len(expected))
 
     @mock.patch('bq_utils.create_table')
     @mock.patch('tools.generate_ext_tables.create_and_populate_source_mapping_table')
@@ -98,19 +102,17 @@ class GenerateExtTablesTest(unittest.TestCase):
         for cdm_table in common.AOU_REQUIRED:
             if cdm_table not in [common.PERSON, common.DEATH, common.FACT_RELATIONSHIP]:
                 query = dict()
-                query[cdr_consts.QUERY] = gen_ext.REPLACE_SRC_QUERY.format(project_id=self.project_id,
-                                                                           dataset_id=self.dataset_id,
-                                                                           combined_dataset_id=self.dataset_id,
-                                                                           mapping_table_id=gen_ext.MAPPING_PREFIX
-                                                                                            +cdm_table,
-                                                                           site_mappings_table_id=gen_ext.SITE_TABLE_ID,
-                                                                           cdm_table_id=cdm_table)
+                query[cdr_consts.QUERY] = gen_ext.REPLACE_SRC_QUERY.format(
+                    project_id=self.project_id,
+                    dataset_id=self.dataset_id,
+                    combined_dataset_id=self.dataset_id,
+                    mapping_table_id=gen_ext.MAPPING_PREFIX + cdm_table,
+                    site_mappings_table_id=gen_ext.SITE_TABLE_ID,
+                    cdm_table_id=cdm_table
+                )
                 query[cdr_consts.DESTINATION_TABLE] = cdm_table + gen_ext.EXT_TABLE_SUFFIX
                 query[cdr_consts.DESTINATION_DATASET] = self.dataset_id
                 query[cdr_consts.DISPOSITION] = bq_consts.WRITE_EMPTY
                 expected.append(query)
         actual = gen_ext.get_generate_ext_table_queries(self.project_id, self.dataset_id, self.dataset_id)
         self.assertCountEqual(expected, actual)
-
-    def TearDown(self):
-        bq_utils.delete_table(gen_ext.SITE_TABLE_ID, self.bq_dataset_id)
