+++ data_steward/tests/unit_test/data_steward/tools/snapshot_by_query_test.py
@@ -7,7 +7,7 @@ WHITESPACE = '[\t\n\\s]+'
 SPACE = ' '
 
 
-class RetractDataGcsTest(unittest.TestCase):
+class SnapshotByQueryTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         print('**************************************************************')
